requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
lxml==4.3.1
Pillow==5.4.1
python-docx==0.8.10
requests==2.21.0
urllib3==1.24.1
